Split package issue with com.ibm.ejs.j2c - Part 1: stop importing own package from com.ibm.ws.jca.cm bundle
