+++ dev/com.ibm.ws.container.service.compat/src/com/ibm/ws/jca/cm/handle/HandleListInterface.java
@@ -8,13 +8,13 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ejs.j2c;
+package com.ibm.ws.jca.cm.handle;
 
 /**
- * Copied to the com.ibm.ws.jca.cm.handle package in order to avoid causing split packages
- * on com.ibm.ej2.j2c with the com.ibm.ws.jca.cm bundle.
- * TODO see if we can switch over usage throughout all other bundles and
- * completely remove com.ibm.ej2.j2c from com.ibm.ejs.j2c from container.service.compat
+ * TODO implementation needed
  */
-public interface HandleListInterface extends com.ibm.ws.jca.cm.handle.HandleListInterface {
+public interface HandleListInterface {
+    void parkHandle();
+
+    void reAssociate();
 }
