+++ dev/com.ibm.ws.container.service.compat/src/com/ibm/ws/threadContext/ConnectionHandleAccessorImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012 IBM Corporation and others.
+ * Copyright (c) 2012,2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,7 @@
  *******************************************************************************/
 package com.ibm.ws.threadContext;
 
-import com.ibm.ejs.j2c.HandleListInterface;
+import com.ibm.ws.jca.cm.handle.HandleListInterface;
 
 /**
  * Stub until LazyAssociatableConnectionManager is designed with J2C.
