+++ dev/com.ibm.ws.ejbcontainer.core/src/com/ibm/ejs/container/EJBThreadData.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2010, 2016 IBM Corporation and others.
+ * Copyright (c) 2010, 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -57,7 +57,7 @@ public class EJBThreadData
      */
     private final FastStack<Object> ivClassLoaderStack = new FastStack<Object>();
 
-    final ThreadContext<HandleListInterface> ivHandleListContext =
+    final ThreadContext<com.ibm.ws.jca.cm.handle.HandleListInterface> ivHandleListContext =
                     getCurrentThreadContext(ConnectionHandleAccessorImpl.getConnectionHandleAccessor().getThreadContext()); // d662032
 
     /**
