+++ dev/com.ibm.ws.jca.cm/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2017,2019 IBM Corporation and others.
+# Copyright (c) 2017,2020 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
@@ -25,7 +25,9 @@ Export-Package: \
     com.ibm.wsspi.zos.tx \
 
 Import-Package: \
+    com.ibm.ws.jca.cm.handle, \
     com.ibm.ws.util;version="[1.0,2)", \
+    !com.ibm.ejs.j2c, \
     !*.internal.*, *
 
 Private-Package: \
